create API authorization using JWT

+++ controllers/api/v1/post_api.js
@@ -22,11 +22,17 @@ module.exports.index = async function(req,res){
 module.exports.destroy = async function(req,res){
     try{
     let post = await Post.findById(req.params.id);
+    if(post.user == req.user.id){
             post.remove();
             Comment.deleteMany({ post: req.params.id });
             return res.status(200).json({
                 message:"post and associated comments deleted"
             });
+    }else{
+        return res.status(401).json({
+            message:"you cannot delete this post"
+        });
+    }
     }catch(err){
         console.log("*********",err);
         return res.status(200).json({
